[tools]Organize the tools directory structure.
