+++ tools/proj_gen/codelite.py
