+++ tools/proj_gen/codelite_template.project
