+++ tools/proj_gen/codelite_template.workspace
