+++ tools/proj_gen/template.cbp
