+++ tools/proj_gen/vs2012.py
