+++ tools/utils/as.sh
